Por si acaso

+++ SIGEBI.Application/Repositories/Configuration/ILibrosRepository.cs
@@ -5,7 +5,7 @@ namespace SIGEBI.Application.Repositories.Configuration
 {
     public interface ILibrosRepository : IBaseRepository<Libro>
     {
-        Task<Libro?> GetLibroById(int id);
+        Task<Libro?> GetLibroById(Int64 id);
         Task<List<Libro>> GetLibroByTitulo(string titulo);
         Task<List<Libro>> GetLibroByAutor(string autor);
         Task<List<Libro>> GetLibroByEditorial(string editorial);
